+++ src/Job/Schedule/Range.php
@@ -6,6 +6,7 @@ namespace Innmind\Cron\Job\Schedule;
 use Innmind\Immutable\Str;
 
 /**
+ * @internal
  * @psalm-immutable
  */
 final class Range