+++ tests/Job/Schedule/MonthsTest.php
@@ -126,7 +126,11 @@ class MonthsTest extends TestCase
     public function testReturnNothingWhenUsingRandomString()
     {
         $this
-            ->forAll(Set::strings()->filter(static fn($value) => !\is_numeric($value)))
+            ->forAll(
+                Set::strings()
+                    ->filter(static fn($value) => !\is_numeric($value))
+                    ->filter(static fn($value) => $value !== '*'),
+            )
             ->then(function($value) {
                 $schedule = Months::maybe($value)->match(
                     static fn($schedule) => $schedule,