+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Changed
 
 - Requires `innmind/foundation:~1.5`
+- `Innmind\Cron\Crontab::_invoke()` now returns an `Innmind\Immutable\Attempt<Innmind\Immutable\SideEffect>`
+- `Innmind\Cron\Read::_invoke()` now returns an `Innmind\Immutable\Attempt`
+- `Innmind\Cron\Schedule\Range` is now declared internal
 
 ## 3.2.0 - 2023-09-23